fix(prisma): prevent haravan_ordere_id type mismatch w/ prod DB

+++ src/controllers/payment/manual-pay/manual-payment-controller.js
@@ -45,7 +45,7 @@ export default class ManualPaymentsController {
 
     const haravanOrderIdStr = body.haravan_order_id;
     if (haravanOrderIdStr != null && haravanOrderIdStr !== "") {
-      paymentData.haravan_order_id = BigInt(haravanOrderIdStr);
+      paymentData.haravan_order_id = Int(haravanOrderIdStr);
     } else {
       delete paymentData.haravan_order_id;
     }
